requirements.txt
flask>=3.1,<4
sqlalchemy>=2.0,<3
pydantic>=2.0,<3
pytest>=8.0,<9
rich>=13.7,<14
textual>=0.47,<1
pyyaml>=6.0,<7
requests>=2.31,<3
